feat: use GitHub packaging for built-in tools definitions
The web_browsing tool isn't working right now because of a bug in
gptscript, so that tool is skipped here.

Additionally, the code_interpreter tool is not ready to be in prime
time, so it will be completed and moved to gptscript-ai at a later time.

Finally, this change does not include the ability to configure the URLs
for the built-in tools.

Signed-off-by: Donnie Adams <donnie@acorn.io>