fix: trim the gptscript_ identifier when running tool with gptscript
There is a difference between the OpenAI built-in tools (like
code_interpreter) and the clicky-chats built-in tools (like
web_browsing): the former will comeback from chat completion without the
prefix and the latter will have the prefix. To simplify, we now trim the
prefix before looking for the tool program to run.

Signed-off-by: Donnie Adams <donnie@acorn.io>